Moving handlers into websockets.js
So, basically, the problem here is that you can't be sure the players
are synced up. One player can be on a different screen. So relying on
passage ordering to register the handlers is a little dicey.

So, let's just not do that. The MAJOR minus is that it decouples the
handler pretty decisively from the passage, which makes the coding
gnarly.